+++ uniconfig/python/frinx_worker/uniconfig/device_discovery.py
@@ -1,6 +1,5 @@
 from ipaddress import IPv4Address
 from ipaddress import IPv6Address
-from typing import Any
 
 import pydantic
 import requests
@@ -17,14 +16,14 @@ from frinx.common.worker.task_def import TaskOutput
 from frinx.common.worker.task_result import TaskResult
 from frinx_api.uniconfig import OperationsDiscoverPostResponse
 from frinx_api.uniconfig.device.discovery.discover import Address
+from frinx_api.uniconfig.device.discovery.discover import DeviceDiscoveryTypeOfAddressModel
+from frinx_api.uniconfig.device.discovery.discover import DeviceDiscoveryTypeOfAddressModel1
+from frinx_api.uniconfig.device.discovery.discover import DeviceDiscoveryTypeOfAddressModel2
+from frinx_api.uniconfig.device.discovery.discover import DeviceDiscoveryTypeOfAddressModel3
+from frinx_api.uniconfig.device.discovery.discover import DeviceDiscoveryTypeOfPortModel
+from frinx_api.uniconfig.device.discovery.discover import DeviceDiscoveryTypeOfPortModel2
 from frinx_api.uniconfig.device.discovery.discover import Input
 from frinx_api.uniconfig.device.discovery.discover import TcpPortItem
-from frinx_api.uniconfig.device.discovery.discover import TypeOfAddressModel
-from frinx_api.uniconfig.device.discovery.discover import TypeOfAddressModel1
-from frinx_api.uniconfig.device.discovery.discover import TypeOfAddressModel2
-from frinx_api.uniconfig.device.discovery.discover import TypeOfAddressModel3
-from frinx_api.uniconfig.device.discovery.discover import TypeOfPortModel
-from frinx_api.uniconfig.device.discovery.discover import TypeOfPortModel2
 from frinx_api.uniconfig.device.discovery.discover import UdpPortItem
 from frinx_api.uniconfig.rest_api import Discover
 from pydantic import Field
@@ -35,37 +34,6 @@ from . import class_to_json
 from . import handle_response
 from .util import get_list_of_ip_addresses
 from .util import parse_ranges
-
-
-def _unwrap_data(discovery_input: Input) -> dict[str, Any]:
-    """
-    Unwrapping is necessary because of the input now containing choice wrapper nodes which cannot be parsed by
-    UniConfig yet. https://frinxhelpdesk.atlassian.net/browse/UNIC-1764
-    :param discovery_input: Device Discovery input.
-    :return: Unwrapped data as dict.
-    """
-    tcp_port: list[dict[str, Any]] = []
-    udp_port: list[dict[str, Any]] = []
-    address: list[dict[str, Any]] = []
-    if discovery_input.tcp_port is not None:
-        for tcp_port_item in discovery_input.tcp_port:
-            if tcp_port_item.type_of_port is not None:
-                tcp_port.append(tcp_port_item.type_of_port.model_dump())
-    if discovery_input.udp_port is not None:
-        for udp_port_item in discovery_input.udp_port:
-            if udp_port_item.type_of_port is not None:
-                udp_port.append(udp_port_item.type_of_port.model_dump())
-    if discovery_input.address is not None:
-        for address_item in discovery_input.address:
-            if address_item.type_of_address is not None:
-                address.append(address_item.type_of_address.model_dump())
-    return {
-        "input": {
-            "address": None if not address else address,
-            "tcp_port": None if not tcp_port else tcp_port,
-            "udp_port": None if not udp_port else udp_port
-        }
-    }
 
 
 class DeviceDiscoveryWorkers(ServiceWorkersImpl):
@@ -90,27 +58,27 @@ class DeviceDiscoveryWorkers(ServiceWorkersImpl):
                 if len(ip_list) == 1:
                     if "/" in ip_list[0]:
                         address = Address(
-                            type_of_address=TypeOfAddressModel3(
+                            type_of_address=DeviceDiscoveryTypeOfAddressModel3(
                                 network=str(IPvAnyNetwork(ip_list[0]))
                             )
                         )
                     else:
                         address = Address(
-                            type_of_address=TypeOfAddressModel(
+                            type_of_address=DeviceDiscoveryTypeOfAddressModel(
                                 ip_address=str(IPvAnyAddress(ip_list[0]))
                             )
                         )
                 else:
                     try:
                         address = Address(
-                            type_of_address=TypeOfAddressModel1(
+                            type_of_address=DeviceDiscoveryTypeOfAddressModel1(
                                 start_ipv4_address=str(IPv4Address(ip_list[0])),
                                 end_ipv4_address=str(IPv4Address(ip_list[1]))
                             )
                         )
                     except ValueError:
                         address = Address(
-                            type_of_address=TypeOfAddressModel2(
+                            type_of_address=DeviceDiscoveryTypeOfAddressModel2(
                                 start_ipv6_address=str(IPv6Address(ip_list[0])),
                                 end_ipv6_address=str(IPv6Address(ip_list[1]))
                             )
@@ -122,7 +90,7 @@ class DeviceDiscoveryWorkers(ServiceWorkersImpl):
             def validate_tcp(cls, tcp_port: str) -> list[TcpPortItem] | None:
                 if tcp_port:
                     return [TcpPortItem(
-                                type_of_port=TypeOfPortModel(
+                                type_of_port=DeviceDiscoveryTypeOfPortModel(
                                     port=p
                                 )
                         ) for p in parse_ranges(tcp_port.split(","))]
@@ -133,7 +101,7 @@ class DeviceDiscoveryWorkers(ServiceWorkersImpl):
             def validate_udp(cls, udp_port: str) -> list[UdpPortItem] | None:
                 if udp_port:
                     return [UdpPortItem(
-                        type_of_port=TypeOfPortModel2(
+                        type_of_port=DeviceDiscoveryTypeOfPortModel2(
                                     port=p
                                 )
                         ) for p in parse_ranges(udp_port.split(","))]
@@ -161,9 +129,8 @@ class DeviceDiscoveryWorkers(ServiceWorkersImpl):
                 headers=dict(UNICONFIG_HEADERS),
                 params=UNICONFIG_REQUEST_PARAMS,
                 method=Discover.method,
-                # temporary workaround until UC adds possibility to accept choice nodes
                 data=class_to_json(
-                    _unwrap_data(template),
+                    template,
                 ),
             )