+++ misc/python/kafka/CHANGELOG.md
@@ -9,3 +9,6 @@
 
 # 1.1.0
 - Support Kafka headers on the input of the Kafka_publish task.
+
+# 1.2.0
+- Updated frinx-python-sdk to version 2.2.1 from 2.1.3.